% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet.R
\name{jet_const}
\alias{jet_const}
\title{A Constant Jet}
\usage{
jet_const(v, lay)
}
\arguments{
\item{v}{The value.}

\item{lay}{A layout from \code{\link{jet_layout}}; taken from the jet
itself unless given.}
}
\value{
A jet.
}
\description{
A number with every derivative zero.
}
\examples{
lay <- jet_layout(2)
x <- jet_var(1, list(2, 1, 0, 0, 0), lay)
acc <- jet_const(0, lay)   # a running sum starts at a constant
(acc + x * 3)$v

}
\seealso{
\code{\link{jet_var}}, \code{\link{jet_layout}}
}
